refactor(tests) move off CCSC tests to other file - clean up api.test.ts - move all CCSC tests to its own file
